etl_ugc/requirements.txt
clickhouse-driver==0.2.5
kafka-python==2.0.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
tzdata==2022.7
tzlocal==4.2
pydantic==1.9.0
